Conflicts:
	src/my/mymoviesamn/JInternalFrame_Favorites.java
	src/my/mymoviesamn/JInternalFrame_Statistics.java

+++ src/my/mymoviesamn/JInternalFrame_Statistics.java
@@ -217,12 +217,12 @@ public class JInternalFrame_Statistics extends javax.swing.JInternalFrame {
         //Διαπερνάω τη Λίστα 2 των Αγαπημένων Λιστών 
         for (int i = 1; i < flist.size() + 1; i++) {
             //Με ένα for διαβάζω τη Λίστα 1 και την περνάω στον βοηθητικό πίνακα
-            for (Movie movie : movies) {
-                //Βρίσκω από την κάθε Αγαπημένη Λίστα την Ταινία με την υψηλότερη Βαθμολογία
-                if (i == movie.getFavoriteListId().getId()) {
-                    String rating = Double.toString(movie.getRating());
-                    model.addRow(new String[]{movie.getTitle(), rating});
-                    break;
+            for (Movie movie:movies){
+            //Βρίσκω από την κάθε Αγαπημένη Λίστα την Ταινία με την υψηλότερη Βαθμολογία
+            if (i== movie.getFavoriteListId().getId()){
+                String rating = Float.toString(movie.getRating());
+                model.addRow(new String[]{movie.getTitle(),rating});
+                break;
                 }
             }
         }
@@ -247,8 +247,8 @@ public class JInternalFrame_Statistics extends javax.swing.JInternalFrame {
         model.setColumnIdentifiers(new String[]{"Ταινία", "Βαθμολογία"});
         //Με ένα for διαβάζω τη Λίστα και την περνάω στον βοηθητικό πίνακα
         movies.forEach((movie) -> {
-            //αφού μετατρέψω τη βαθμολογία από double σε string
-            String rating = Double.toString(movie.getRating());
+            //αφού μετατρέψω τη βαθμολογία από float σε string
+            String rating = Float.toString(movie.getRating());
             //Προσθέτω μία μία γραμμή της Λίστας στον Πίνακα
             model.addRow(new String[]{movie.getTitle(), rating});
         });
